Current implementations of Modal deployment
Since we wish to have our model deployed in a way that does not require the TA to run computations on his own computer, I have started testing with model deployment on Modal. Still in the works since I need to implement it properly into the frontend and connect it to the flask but for now this is a file that sets up the model with all of its dependencies.